Restructure test file for better review
